move CycloptsPanel to it's own module.
